rename .rs files
move the const definitions in the current virtio_gpu.rs to protocol.rs
rename vhu_gpu.rs to device.rs
rename virt_gpu.rs to virtio_gpu.rs
