+++ staging/vhost-device-gpu/src/main.rs
@@ -13,8 +13,8 @@ use thiserror::Error as ThisError;
 use vhost_user_backend::VhostUserDaemon;
 use vm_memory::{GuestMemoryAtomic, GuestMemoryMmap};
 
-use vhost_device_gpu::vhu_gpu;
-use vhost_device_gpu::vhu_gpu::VhostUserGpuBackend;
+use vhost_device_gpu::device;
+use vhost_device_gpu::device::VhostUserGpuBackend;
 use vhost_device_gpu::GpuConfig;
 
 type Result<T> = std::result::Result<T, Error>;
@@ -22,7 +22,7 @@ type Result<T> = std::result::Result<T, Error>;
 #[derive(Debug, ThisError)]
 pub(crate) enum Error {
     #[error("Could not create backend: {0}")]
-    CouldNotCreateBackend(vhu_gpu::Error),
+    CouldNotCreateBackend(device::Error),
     #[error("Could not create daemon: {0}")]
     CouldNotCreateDaemon(vhost_user_backend::Error),
     #[error("Fatal error: {0}")]
